+++ src/pages/admin/application/[id].jsx
@@ -5,10 +5,13 @@ export default function AdminApplicationDetailPage() {
     <>
       <Head>
         <title>신청 관리 디테일 페이지</title>
-        <meta name="description" content="관리자에게 보여지는 신청 관리 디테일 페이지입니다." />
+        <meta
+          name="description"
+          content="관리자에게 보여지는 신청 관리 디테일 페이지입니다."
+        />
       </Head>
       <div>
-      <h1>신청 관리 디테일 페이지 내용</h1>
+        <h1>신청 관리 디테일 페이지 내용</h1>
       </div>
     </>
   );
