+++ src/pages/auth/login.jsx
@@ -18,7 +18,7 @@ export default function LoginPage() {
       <p className={styles.text}>
         회원이 아니신가요?
         <Link href="/auth/sign-up" className={styles.link}>
-          로그인
+          회원가입하기
         </Link>
       </p>
     </>
