+++ src/pages/auth/sign-up.jsx
@@ -15,7 +15,7 @@ export default function SignUpPage() {
       <p className={styles.text}>
         회원이신가요?
         <Link href="/auth/login" className={styles.link}>
-          가입하기
+          로그인하기
         </Link>
       </p>
     </>
